Force Django to import all necessary modules before first request
When running on Vagrant, the interop repo (which contains the server
script) is shared with the guest via the synced_folders Vagrant config.
Unfortunately, this folder can be very slow, so the first request can
have significant latency, which is occasionally long enough to make
requests time out, and tests fail.

We already tell Apache to do some preloading, which preloads most of
Django, but Django itself lazily loads most of the modules needed to
service requests only when the first request arrives.

By sending a dummy request in the WSGI setup, we ensure that the first
real request has low latency.

Fixes #8